pyproject.toml
[tool.poetry]
name = "cookiecutter"
version = "0.1.0"
description = ""
authors = ["Jhonatan da Silva"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
drf-spectacular = "^0.27.1"
redis = "^5.0.2"
django-redis = "^5.4.0"
psycopg2 = "^2.9.9"
ipdb = "^0.13.13"
factory-boy = "^3.3.0"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-django = "^4.8.0"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
